organise auth screens into one folder
